requirements.txt
svg.py==1.10.0
requests>=2.31.0
